requirements.txt
flask
langchain
langchain-together
python-dotenv
chroma
sentence-transformers
chromadb
